+++ scraperx/extract.py
@@ -325,8 +325,13 @@ class Extract(ABC):
         for qa_field, qa_rule in qa_rules.items():
             # Make sure key exists
             if qa_field not in result:
-                raise QAValueError((f"Field {qa_field} is missing from data"
-                                    f" at result {idx}"))
+                if 'default' in qa_rule:
+                    result[qa_field] = qa_rule['default']
+                    # No need to check other params since this was user set
+                    continue
+                else:
+                    raise QAValueError((f"Field {qa_field} is missing from data"
+                                        f" at result {idx}"))
             # Check required
             if result[qa_field] is None:
                 if qa_rule.get('required', False) is True: